+++ changelog.md
@@ -7,3 +7,14 @@
     - Change size and fix speed
     - Change moving logic to fix the bug where sometimes very fast squares can pass through other ones
     - Increase creation chance substantially
+
+# Update 1.2: The age update
+## New Features and changes
+    - Added life spans for squares! Now squares can slowly die. If they bump into each other, they will die a bit faster
+    - Added a pause menu, accessible by pressing p or esc.
+    - Added a restart button and 2 debug buttons. I wonder what they do?
+    - Another buff to spawn chance of squares, and a small buff to the number of squares on screen
+    - Squares now slowly turn white as they become "old"
+## Fixes
+    - Fix the issue where vector keeps being 'flee' when pausing inside an ROE
+    - Fix ROE effect to be more appealing